Internal: Refactoring callbacks code (cont.)
Moving related code to callbacks.py
